formatting, renamed error_modal to modal
