+++ src/app/components/mod.rs
@@ -1,4 +1,4 @@
 pub(crate) mod download_item;
-pub(crate) mod error_modal;
 pub(crate) mod loading_wheel;
+pub(crate) mod modal;
 pub(crate) mod nav_sidebar;
